Revert "Feature/vrtool 28 read vrtool config from json"
