+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "vrtool"
+name = "veiligheidsrendement tool"
 version = "0.0.1"
 description = ""
 authors = ["Wouter Jan Klerk <wouterjan.klerk@deltares.nl>"]
