--- src/run_workflows/vrtool_plot_mode.py
@@ -1,10 +1,0 @@
-from enum import Enum
-
-
-class VrToolPlotMode(Enum):
-    """
-    Definition of the different types of plots supported during a run of a `VrtoolRunProtocol` instance.
-    """
-
-    STANDARD = 0
-    EXTENSIVE = 1
