--- tests/run_workflows/measures_workflow/test_results_measures.py
@@ -1,11 +1,0 @@
-from src.run_workflows.measures_workflow.results_measures import ResultsMeasures
-from src.run_workflows.vrtool_run_result_protocol import VrToolRunResultProtocol
-
-
-class TestResultsMeasures:
-    def test_init(self):
-        _results = ResultsMeasures()
-
-        assert isinstance(_results, ResultsMeasures)
-        assert isinstance(_results, VrToolRunResultProtocol)
-        assert _results.solutions_dict == {}
