--- tests/run_workflows/safety_workflow/test_results_safety_assessment.py
@@ -1,12 +1,0 @@
-from src.run_workflows.safety_workflow.results_safety_assessment import (
-    ResultsSafetyAssessment,
-)
-from src.run_workflows.vrtool_run_result_protocol import VrToolRunResultProtocol
-
-
-class TestResultsSafetyAssessment:
-    def test_init(self):
-        _results = ResultsSafetyAssessment()
-
-        assert isinstance(_results, ResultsSafetyAssessment)
-        assert isinstance(_results, VrToolRunResultProtocol)
