--- tests/test_data/integrated_SAFE_16-3_small/custom_config.json
@@ -1,3 +1,0 @@
-{
-    "traject": "16-3"
-}
