--- tests/test_data/vrtool_config/custom_config.json
@@ -1,3 +1,0 @@
-{
-    "traject": "MyCustomTraject"
-}
